Fix feature policy in isolated sandboxes.
This CL adds a new member to the ParsedFeaturePolicyDeclaration, which
indicates that the iframe policy is expected to apply to the sandboxed
frame. This is triggered with an iframe of the form

<iframe sandbox allow="feature">

or

<iframe sandbox allow="feature src">

Because the opaque origin of the frame is not known when the
HTMLFrameOwnerElement builds the container policy, this flag is checked
when building the feature policy in the new frame, and ensures that the
new feature policy will allow the feature in that origin.

Bug: 794631, 690520
Change-Id: Ie18b9bc3c36be6550baf5a03e355871b9589fd40