Add memory of last SDP offer/answer created
This adds internal slots "LastOffer" and "LastAnswer", and uses
those to check that SetLocalDescription uses an unchanged SDP offer/answer.

Because modification of SDP is suspected to be used in a number of places,
this change only rejects SDP with modified fingerprints (which would fail
anyway), but merely counts the usage for other modifications.

Bug: chromium:823036
Change-Id: I0c978d5ff3e63b0afab3ec02334c57a5aaa94cdd